**Init `Minimum cards due` preference individually
Use simple summary provider to improve summary setting reliability

And remove `initAllPreferences` as that is no longer used
