+++ AnkiDroid/src/main/java/com/ichi2/anki/preferences/NotificationsSettingsFragment.kt
@@ -15,9 +15,18 @@
  */
 package com.ichi2.anki.preferences
 
+import android.app.AlarmManager
+import android.content.Context.ALARM_SERVICE
+import android.content.Intent
+import androidx.preference.ListPreference
 import androidx.preference.SwitchPreference
+import com.ichi2.anki.Preferences
 import com.ichi2.anki.Preferences.SpecificSettingsFragment
 import com.ichi2.anki.R
+import com.ichi2.anki.services.BootService.Companion.scheduleNotification
+import com.ichi2.anki.services.NotificationService
+import com.ichi2.compat.CompatHelper
+import com.ichi2.libanki.utils.TimeManager
 import com.ichi2.utils.AdaptionUtil.isRestrictedLearningDevice
 
 /**
@@ -35,5 +44,36 @@ class NotificationsSettingsFragment : SpecificSettingsFragment() {
             preferenceScreen.removePreference(requirePreference<SwitchPreference>(R.string.pref_notifications_vibrate_key))
             preferenceScreen.removePreference(requirePreference<SwitchPreference>(R.string.pref_notifications_blink_key))
         }
+        // Minimum cards due
+        // The number of cards that should be due today in a deck to justify adding a notification.
+        requirePreference<ListPreference>(R.string.pref_notifications_minimum_cards_due_key).apply {
+            updateNotificationPreference(this)
+            setOnPreferenceChangeListener { preference, newValue ->
+                updateNotificationPreference(preference as ListPreference)
+                if ((newValue as String).toInt() < Preferences.PENDING_NOTIFICATIONS_ONLY) {
+                    scheduleNotification(TimeManager.time, requireContext())
+                } else {
+                    val intent = CompatHelper.compat.getImmutableBroadcastIntent(
+                        requireContext(), 0,
+                        Intent(requireContext(), NotificationService::class.java), 0
+                    )
+                    val alarmManager = requireActivity().getSystemService(ALARM_SERVICE) as AlarmManager
+                    alarmManager.cancel(intent)
+                }
+                true
+            }
+        }
+    }
+
+    private fun updateNotificationPreference(listPreference: ListPreference) {
+        val entries = listPreference.entries
+        val values = listPreference.entryValues
+        for (i in entries.indices) {
+            val value = values[i].toString().toInt()
+            if (entries[i].toString().contains("%d")) {
+                entries[i] = String.format(entries[i].toString(), value)
+            }
+        }
+        listPreference.entries = entries
     }
 }
